+++ src/main/kotlin/com/zp4rker/dsc/core/extenstions/LoggerExtensions.kt
@@ -1,0 +1,11 @@
+package com.zp4rker.dsc.core.extenstions
+
+import org.slf4j.Logger
+
+/**
+ * @author zp4rker
+ */
+
+fun Logger.separator() {
+    info("=".repeat(40))
+}