+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v1.0.0-release
-- Re-published as `com.zp4rker.dsc.disbot`
+- Re-published as `com.zp4rker.dsc:disbot-core`
 - Started a changelog
 - Started a proper README